Respect the NO_COLOR standard (#44)
* feat(control): respect NO_COLOR behavior

Similarly to the behavior for CLICOLOR/CLICOLOR_FORCE, this adds support
for the NO_COLOR spec (https://no-color.org/).

Related to #2.

* refactor(control): make NO_COLOR as an alias of CLICOLOR_FORCE == 0

Rather than adding a new field to the struct, this treat `NO_COLOR` as
an alias for `CLICOLOR_FORCE == 0`. However, this still gives priority
to an explicitly set `CLICOLOR_FORCE`. That is, if `NO_COLOR == 1` and
`CLICOLOR_FORCE == 1`, color WILL be output.

See #44.

* docs(readme): add lukehsiao as contributor

* refactor(control): resolve NO_COLOR and CLICOLOR_FORCE in private function

Moving to a private function also allows for easier unit testing. In
addition, this commit adds a documentation comment for the intended
behavior of ShouldColorize::from_env().